Show selected variations within product row
